Introduce lsp-connect to allow users to handle raw LSP responses
The parent commit allows to retrieve raw JSON responses from the
language server. However, it still requires a sizeable shell script
that relies on knowing some parts of kak-lsp's implementation.

Introduce lsp-connect, which can call selected lsp-* commands,
and forward the server's JSON response to a user-defined handler.
The advantage of making this an editor command (instead of a
commandline option) is that we get the editor context for free.

lsp-connect only supports selected LSP commands that are deemed
(remotely) useful. For example, lsp-incoming-calls would not be useful
yet, because it requires multiple requests to return meaningful data.
In future we should probably make lsp-incoming-calls return the
location list it would render otherwise.

lsp-connect is blocking (but does not have the "-sync" suffix) because it
is implemented by setting a global option, so making it async would create
race conditions. In my testing, the blocking requests are fast enough.

Observe that lsp-connect might allow us to reimplement
lsp{,-hover}-{next,previous}-{symbol,function} without special
support in Rust, since these functions are essentially UI on top
of lsp-document-symbol. (However, the problem with LSP vs. Kakoune
coordinates remains.)

Add a hack to write JSON for diagnostics; this is not exactly the
format sent by the server sends but the information is the same.

Closes #580

For experimenting, this handler could help:

	define-command cat-debug -params 1 %{
	    evaluate-commands "echo -debug %%file{%arg{1}}"
	    buffer *debug*
	}

Here is a real-world usage example: echo the name of the top-level
function (or class etc.) around the cursor. VSCode does something
similar, except they show *all* parent symbols. (We should totally
copy that feature.)

	map global lsp ? %{: lsp-connect echo-current-toplevel-symbol lsp-document-symbol<ret>}

	define-command echo-current-toplevel-symbol -params 1 %{
	    evaluate-commands %sh{
	        python <"$1" -c "
	import json, sys
	SQ = '\''
	quote = lambda s: SQ + s.replace(SQ, SQ*2) + SQ
	symbols = json.load(sys.stdin)['result']
	# Symbols could be a flat list of symbols, or a tree!
	for i, symbol in enumerate(symbols):
	    if symbol['range']['start']['line']+1 > $kak_cursor_line:
	        if i > 0:
	            symbol = symbols[i-1]
	            break
	print('echo', quote(symbol['name']))"
	    }
	}