Avoid shell calls when applying text edits
As reported on IRC, lsp-formatting is fairly slow at formatting many
lines with PHP's intelephense language server.

This happens because the language server sends many fine
grained text edits, which we apply with lsp-replace-selection or
lsp-insert-selection, both of which shell out to mktemp and cat.

I experimented with this line:

	if (1) ; else ;

which intelephense reformats to

	if (1);
	else;

To reformat 100 of these, we need around 600 calls to
lsp-replace-selection, amounting to 1200 shell calls.  Curiously,
the shell calls are completely unnecessary.  We can simply use insert
mode to perform text edits; we only need to escape keys before passing
them to execute-keys. Do that.
On the reporter's system (with a different input), this makes us
go down from 3m to ~1s.  On my system it's 1 second -> 0.5 seconds
(debug build).

While at it, collapse all edits into a single exec(ute-keys) statement.
This doesn't make a measurable difference system. but perhaps it's
more elegant? Also a bit harder to read, but let's keep it for now.

Alternative/future solution: don't send text edits to Kakoune but apply
all text edits on the server side and just replace the entire buffer.