Avoid creating implicit scratch buffer after deleting temp buffer
At load time, we create a temporary buffer to detect if the system's
wcwidth(3) is modern enough, and we can safely use emoji.  We delete
this buffer immediately after use.  If this happens inside kakrc
there is no buffer left (besides *debug*), so Kakoune will create a
*scratch* buffer, even if the user supplied a file argument.

Fix this by switching to the *debug* buffer before deleting our scratch
buffer. This means that Kakoune won't create the *scratch* buffer.

To make this also work when loading lsp.kak after loading kakrc, run
the detection commands in a draft context if possible.  The draft
context fails when inside kakrc with "no selections available",
so let's fall back to not using -draft.

Fixes #593