Support proposed <a-x> change
The breaking-cleanups branch in the Kakoune repo changes the behavior
of <a-x>, and instead gives x that behavior. Let's use x, so kak-lsp
works with that branch. Make sure it keeps working with the current
semantics by reducing the selection to the cursor before running x
(else x might go to the next line).