Introduce text objects for selecting functions and other symbols
lsp-{next,previous}-{function,symbol} - bound to "[]{}90()" in the
lsp user mode have proven useful to my workflow. Let's expand on
these ideas with an approach that is more native to Kakoune: text
objects.
Introduce lsp-text-object, which moves similar to ]p, ]( and friends,
except it matches LSP symbol ranges.

New object mode mappings were chosen to not collide with any of
another plugin, https://github.com/Delapouite/kakoune-text-objects

"<a-a>" is a convenience alias for "a", since "<a-a>a" is awkward
to type.

The behavior is a bit different from most text objects:
Similar to ]), ]a selects only the symbol range, not the text in
between the cursor.
OTOH, }a behaves more like normal text objects: it tries to include
the entire current selection.

The surrounding modes <a-a> and <a-i> only select nodes that surround
the cursor. This could be relaxed in future.

The implementation is pretty hacky (it doesn't convert LSP coordinates
properly) but we can fix that later.