Allow *-sync requests if the controller is down
Synchronous requests are implemmented by writing the response to
meta.fifo. If there is no controller process (kak-lsp server), we
first need to spawn it and perform initialization, which requires some
roundtrips to the editor. Since the Kakoune thread is blocked reading
meta.fifo, this is a deadlock. Avoid this by sending initialization
commands via $kak_command_fifo.  In future we should rely on this
feature more and probably use $kak_response_fifo for synchronization.

Steps to reproduce the fixed issue:
1. set a low server timeout in kak-lsp.toml (e.g. 5 seconds)
2. edit a file that has wrong formatting (so lsp-formatting-sync will correct it)
3. run lsp-enable and wait until the controller terminates
4. run lsp-formatting-sync. This used to show a popup

	language server is not running, cancelling blocking request

but works now; it starts a fresh controller and formats the file.