Workaround to keep selections when servers send whole-buffer text edits
When we apply whole-buffer text edits we move the cursor to the end
of the buffer. Let's avoid this by computing a smaller edit script
ourselves.

This is better than the previous attempt (40d3f78f (Preserve cursor
positions again when applying text edits, 2022-04-02)) because it
allows to maintain selections better when working with language
servers that send smaller (as in affecting fewer lines) edit scripts,
for example for test/rust-analyzer-code-actions.sh.

Our diff is line-based. We should probably remove this restriction,
at least for small inputs.