Also send server-specific settings from kak-lsp.toml in workspace/didChangeConfiguration
Server specific settings can be configured in two ways
1. statically in kak-lsp.toml
2. dynamically via the option lsp_config (or one of the deprecated options)

On "initialize", we send either the dynamic settings if
defined, or else the static ones.  Whenever dynamic settings
change (or when we open a new buffer), we send the notification
"workspace/didChangeConfiguration" with the dynamic settings. This
should be enough because the server already knows the static ones from
"initialize". Sadly, as reported in #611, pylsp completely ignores
intialization options. This feels like a bug in pylsp, but LSP does not
prescribe the sensible behavior, so we probably have to live with it.

Let workspace/didChangeConfiguration fall back to static configuration
if neither lsp_config nor the legacy option are defined.

This is an ugly workaround, because now, by default, we send
initialization options twice. We should ask pylsp to fix that.

Closes #611