Support resolving completion items' "detail" and "documentation" fields
As reported in [1], typescript-language-server sends completions
without documentation, and only provides documentation after we send
an extra completionItem/resolve request.

Today, we only send this request when a completion is accepted or
rejected (to apply any additionalTextEdits). Completion documentation
however needs to be resolved as soon as the user selects a completion.
Do that (if the server is a resolveProvider), except for completions
that already have documentation (like rust-analyzer's), where we assume
that it's not gonna change.

[1]: https://discuss.kakoune.com/t/unable-to-verify-if-kak-lsp-is-working-mac/2051/16