lsp-hover: tell user if code actions are available
But not in lsp-hover-buffer because that seems irrelevant.

A diagnostics is often accompanied by a code action ("fix available")
but users might not know that.  Code action availability is signaled
by the lightbulb indicator in the modeline, but new users might not
know how to use that.  Let's add that info to the hover info box.

Unfortunately this only works if lsp_auto_show_code_actions is set
to true.  Maybe we should make that the default; I will test that
for a while.