lsp-hover: tell user if code lenses are available
Just like the parent commit, don't add it to lsp-hover-buffer because
there's no good motivation.

Lines with code lenses are flagged with a '>' character in the gutter.
A new user might not know what this means. We don't have a context
menu but lsp-hover is our next best thing. Let's add a line about
code lenses, to make the '>' thing more discoverable.