lsp-hover: limit to 20 lines via lsp_hover_max_lines
Some language servers (rust-analyzer) send lots of documentation.
The info box is not scrollable, so it can occupy an entire screen.
This is particularly bad since it overshadows diagnostics, and the
new hints about code lenses or code actions. This information is at
the end of the hover info box but it's probably more important than
documentation, especially since documentation can also be read with
gd or lsp-hover-buffer (which we point out in the hover message
if truncated).

Let's limit to 20 hard-wrapped lines by default. This may not be low
enough if the server sends very long (soft-wrapped) lines, but it
seems to work with rust-analyzer. 20 lines fit in a standard terminal,
so users should always see diagnostics by default.