Use new hook to resolve completion items
This fixes an issue where we would send completionItem/resolve
requests if the user didn't select an item.  This is reproducible on
the grandparent commit by first selecting a completion with <c-n>,
then deselecting it with <c-p> and accepting the completion.  Since the
index was still pointing the first item, we'd try to resolve that.