Naturalize snippet support code
This snippet was vendored from
https://github.com/occivink/kakoune-snippets.  It uses aliases for
Kakoune command ("def" instead of "define-command") while the rest of
our code uses long names.  We left the aliases in the snippet code
to make it easier to compare to upstream. The mix of style causes
confusion, so let's use long names everywhere. It's still fairly easy
to see the differences to upstream.