snippets: simplify detection of when a completion is accepted
Whenever a completion is accepted, the InsertCompletionHide hook
parameter has the inserted range (and is thus non-empty).  OTOH, when
completions are rejected, the parameter will be empty.  Since we only
care about accepted completions, adjust the filtering regex so the
hook doesn't fire when a completion is rejected. Rename the symbols
accordingly.