Correct markup escaping for inlay hints
Inlay hints with backslashes are not rendered in some cases because
they result in an invalid markup string. For example if I hardcode
this value in the inlay hints code:

	let label = r#"123\{456"#;

Then inlay hints will fail to render.  This is because we forgot
to escape backslashes in the markup string. Fix this, and also make
another place use the same escaping style.

Here's a similar example that shows how the two levels of escaping
interact:

	kak -e '
		exec %{%dihello world<esc>}
		declare-option range-specs my_ranges
		add-highlighter global/ replace-ranges my_ranges
		# raw value:                      \{
		#    -> escaped for markup:     \\\{
		#      -> escaped for tuple: \\\\\\{
		set-option global my_ranges %val{timestamp} %<1.5+0|\\\\\\{>
		# renders as hell\{o world
	'