Send documentHighlight/hover/signatureHelp requests only if supported by the server
These requests may be sent
automatically via

	lsp_auto_highlight_references
	lsp-auto-hover-enable
	lsp-auto-signature-help-enable

As reported in [1], if a user enables them globally but some language
doesn't support them, they get noisy error messages in an info box.
Let's just do the smart thing here and never send these requests if
unsupported. That kind of feedback is nice when running the requests
manually but it's not a big loss.

Commit 8660d5ab (Send code action requests only if supported by the
server, 2022-04-29) did the same for code actions.

[1]: https://github.com/kak-lsp/kak-lsp/issues/82#issuecomment-1175230471