Create parent directories when documentChanges request new file
Reproduce by adding "mod aaa;" somewhere in a Rust file.  The code
action that adds "aaa/mod.rs" fails because the directory doesn't
exists. The server has asked us to create that file, so it seems
natural that we are creating the directory as well.

BTW permissions are slightly different: I get 640 and 750 in the files
created by kak-lsp whereas existing source files use 644 755. Anyway,
that's unrelated, if at all an actual issue.

Fixes #627