Display an error when a lsp server goes down
Currently, we ignore a request without visible error when a language
server crashes in the background. This adds an error message for that.

I've managed to get rust-analyzer crash on me every so often, which can
be a bit frustrating because it also reacts slowly to requests shortly
after startup. It's also possible for lsp servers to go down for benign
reasons like linux killing the process in a low memory scenario. This is
also reproducible with e.g. `pkill rust-analyzer`.

This isn't perfect; if I send a request within a second of killing the
lsp server, i.e. while the language server transport layer is shutting
down, there is no error reported. But that's a separate issue.