Don't duplicate cursor after completing labeled snippets
When typing "const_ca", clangd sends

	const_cast<${1:type}>(${0:expression})

where we incorrectly think that the final tabstop (AKA $0) is missing.
Fix this by also recognizing braced tabstops.

---

In general, (interaction with) clangd is not really consistent yet.
They sometimes use $1, so we can add $0 at the end, like for
std::vector::add:

	at(${1:size_type n})$0

but for methods with overloads of different arity, we only get $0:

	resize($0)

so without gross hacks we can't move past the closing paren.