Disable completion snippet placeholders when leaving insert mode
When snippets are enabled, we can type

	openat<c-n>somefd<tab>somefile<tab>O_CREAT,<space>0644

to get

	openat(somefd, somefile, O_CREAT, 0644)

Very fancy.
Sadly, I often can't fill in all arguments without looking
around or fixing some other code.  If I stop somewhere in
the middle, highlighting remains, and can only be cleared
with `lsp-snippets-select-next-placeholders`.  For example
`openat<c-n><esc>` gives

	openat(%(), const char *file, int oflag, ...)

where the tail of arguments is still highlighted.  This is supremely
annoying because I usually don't want to simply resume where I left
of.

Also, the last snippet-placeholder is usually after the closing
parenthesis. Jumping there makes sense when staying in insert mode
throughout but not when when exiting insertion mode halfway through
the arguments.

Let's just hide the placeholder highlighting as soon we exit insert
mode. This should make snippets interfere less with normal editing.

It seems fine to leave the raw placeholder text because it serves
as documentation and other editors like VSCode do the same.

If anyone really cares about this, we can use a separate hook group,
that can be disabled specifically.