Recommend mappings to jump to next/previous diagnostic
Helix provides a default mapping for jumping to next/previous error. It
is pretty useful when you know that the current buffer has errors,
so we should at least have a recommended mapping.

Add recommended mappings to the object mode. This is a bit hacky
because they don't select _until_ the error but move the entire
selection, independent of object mode key. However it has the advantage
that users are already familiar with choosing a direction with [ or ].