lsp-signature-help: make active parameter bold
Fellow editor Helix added signature help support in their latest
release https://helix-editor.com/news/release-22-08-highlights/.
They paint the active parameter in a bold font which seems really
nice. Let's do the same.

This means that the argument to lsp-show-signature-help is a markup
string. This is a breaking change but with low risk and low impact,
so let's just sweep it under the rug.