lsp-definition: select targetSelectionRange, not targetRange
Commit 6a66c1bd (lsp-definition: select full symbol name instead of
placing cursor at symbol start, 2022-09-03) uncovered a place
where we don't follow the LSP spec: on textDocuemnt/definition
we should select targetSelectionRange, not targetRange.
This surfaces for language servers that send the entire range of the
symbol as targetRange (which is perfectly valid); we don't want to
select that.

Closes #643

In future we might want to highlight targetRange (temporarily).