Fix off-by-one error converting at-eol UTF-16 character to byte offset
This is the same as 3bb7fbfd (Fix off-by-one error in line number
conversion, 2021-07-13). I found this when trying to apply a
code action from Deno on this type script file:

	function x(): void {
		1 + 1;
	}

With the next commit, put the cursor on x and run lsp-code-actions.
"Move to a new file" used to fail with "Text edit range points past end of
line.".  This is because it creates a new file and then inserts at line
0 character 0.  Character 0 is definitely valid - it's byte offset 0.