Fake filterText support by adding insertText as snippet
The Julia language server provides completion ξ with filterText
"\xi".  Commit 11776913 (Support LSP filterText in completions,
2021-10-31) added support for this (#444) conditional
on a proposed extension to Kakoune's completion system
(https://github.com/mawww/kakoune/pull/4418).

Turns out this extension may not be needed after all.
We can simply fake filterText support by pretending that the filterText
is the actual completion, and later when the completion is accepted,
replace it with the actual insertText.

This solution has two potential drawbacks:
1. the completion menu shows the filterText, not what's inserted by snippets.
   This is both good and bad. Good because it makes Kakoune's filtering
   predictable. Bad because the inserted text is only disclosed quite late.
   However this is something the language server can fix easily: they
   could simply set the "completionItem.label" property to something
   like "\xi → ξ". The PHP language server does something similar,
   see below.
2. the value is only inserted after the completion menu is closed.

Not sure if these are important?

This fix is disabled for completions that are snippets already.
Not sure if it's easy to support those without dropping information.

Another example: given a fresh Laravel Project with this file:

	use Illuminate\Http\Request;

	Request;

if I type on a new line:

	Reque

PHP language server (intelephense) example:
completion menu shows

	Request [GuzzleHttp\Psr7]               Class

The filterText is "Request" (which is also nice for sorting, because
it's next to all other importable requests).  The insertText is
"Psr7Request". The left part is the label. Observe that it nicely
hints at the fact that insertText != filterText.

When accepting this completion, we get

	 use Illuminate\Http\Request;
	+use GuzzleHttp\Psr7\Request as Psr7Request;

	 Request;
	+Psr7Request

Very nice. Previously, we only got "Psr7Request" in the completion
list, which is not what the server intended.