Introduce lsp-inlay-code-lenses to show code lenses as virtual text
Ideally we want to enable them by default, as intended by LSP. Sadly,
Kakoune's virtual text support does not allow adding a virtual line
(above) so we need to show them on the same line. This causes glitches,
so let's wait with enabling it by default.

Many thanks to Sidharth Kshatriya the initial patch.

We discussed an alternative implementation of merging this with inlay
hints. That would work as well, I'm open to that (might simplify UX).

Closes #623