Suppress "language server not initialized" errors that originate from hooks
We already suppress the "language server not initialized" error for
requests like lsp-highlight-references that are probably sent from
hooks. A few are missing however, for example lsp-signature-help.

Using Matthew Pfeiffer's additions, we can detect whether an LSP
request was made during hook execution. Suppress the error if (and
only if) the request was made in a from hook context. This is the
better criteria because it will show an error if lsp-hover is run
interactively.