lsp-code-actions: search entire line for diagnostics' code actions
It is fairly annoying having to navigate to the precise location of
a diagnostic to fix it (especially given that usually there is just
one diagnostic in a line). Let's fix this by requesting code actions
for the entire line in the common case.

Alternative approach: we could first try with the exact selection, and
only if there are no results expand the scope. It's hacky either way :(