Update patched lsp-types to the latest patch version
Until the lsp-types PR is merged we're patching it.  Since adding
that patch I have updated the patch submitted to lsp-types (only
changing the commit message) but that means that other git clients
won't fetch the old commit.  Let's update it.

	$ git range-diff d7219cfaacfe6b6f8e53eb040171f08dc55e4290...a2091e15d1e0743c37241c9c93ccb8657a928e93
	1:  d7219cf ! 1:  a2091e1 Derive Hash for file change events
	    @@ Metadata
	      ## Commit message ##
	         Derive Hash for file change events

	    +    I want to use [here] for implementing
	    +    "workspace/didChangeWatchedFiles", specifically to deduplicate file
	    +    change events before I send them to the server.
	    +
	    +    [here]: https://github.com/kak-lsp/kak-lsp/pull/650/commits/346123f52fc8a1e450ac522938aac6809f177a0f#diff-1fd48ab302efbb91fe84507671007262db18094467e4ad707d036f60ff22e0eaR92
	    +
	      ## src/lib.rs ##
	     @@ src/lib.rs: pub struct DidChangeWatchedFilesParams {
	      }

(To address any security concerns, the integrity of the build is in
any case guaranteed due to the commit SHA in Cargo.lock.)