Don't crash when resolving completion item with invalid index
We send the resolve requests based on completions inside Kakoune.
I believe there's a race condition where the kak-lsp server updates its
completion list before they are updated in Kakoune's completion pager.
This means that we might see brief periods where we might send a
stale index.  An out-of-bounds index has been sighted in the wild
(#654) so let's avoid crashing in this case.

This is a quick-and-dirty workaround. A better workaround would be to
add a timestamp to the completion list, so we can reject all stale
resolve requests, not just the ones that are OOB.  It is probably
not worth the effort to try to fix the race condition.