Sort code action completions by importance
This is inspired by a recent change in [Helix] that fixes sorting of
code actions.
We have the same problem; we use ":prompt -shell-script-candidates"
to show code actions.
For example, on this Rust file:

	fn main() {
	    let f: FnOnce(HashMap<i32, i32>);
	}

with the cursor on "HashMap", a ":lsp-code-actions" will use ":prompt"
to offer two code actions from rust-analyzer:

	Extract type as type alias"
	Import `std::collections::HashMap`

The first one is a refactoring and the second one is a quickfix.

A proposed Kakoune patch preserves completion input order (if the
query is empty). Use this to fix the order of above code actions.

[Helix]: https://github.com/helix-editor/helix/pull/4134