Fix "kak-lsp --request" races when spinning up new server
If "kak-lsp --request" fails to connect to a running server, it spins
up a new one.
As reported in #654, this fails intermittently. The error is

	Failed to daemonize process: LockPidfile(11)

where 11 means EAGAIN, which comes from "Daemonize::new()",
specifically

	let fd = open(path_c.as_ptr(), libc::O_WRONLY | libc::O_CREAT, 0o666);
	tryret!(
		flock(fd, LOCK_EX | LOCK_NB),
		Ok(fd),
		DaemonizeError::LockPidfile
	)

This happens when there are two "kak-lsp --request" processes racing
against each other. First both fail to connect to a server, hence
both try to start a new one. The second one fails because the PID
file is already locked.  Hence this is a time-of-check to time-of-use
(TOCTOU) race condition.

As a result, the process that loses the race is dropped on the floor.
Let's fix this by making it go back to try to connect to the server.
This is done in a cheesy loop, so let's add a sleep to reduce CPU load.

To spin up a server, "kak-lsp --request" actually spawns a "kak-lsp
--daemonize" child process. I do not know why this is necessary but
let's leave it for now.

Closes #654