lsp-hover-buffer: stop using fifo, to avoid races
This request uses a fifo for sending the rendered hover buffer to the
editor. Since we already have rendered everything, there is no point
to going the fifo route. Let's include the rendering in the command
we send back to the editor.

This fixes a race condition with the fifo logic. I could reproduce
the problem but I didn't figure out the root cause.

Closes #664