Make sure to re-request inlay hints and semantic tokens if they didn't succeed
Commit 9621c99d (Send inlay hints and semantic tokens requests only
after buffer changes, 2023-01-14) was an overzealous optimization.
If a request failed, we would still mark it as cached.  Let's do
that only after the request suceeds.  This means we will again send
redundant requests for a brief period but that's okay. We can possibly
improve this once we support cancellation of stale requests (#666).

Closes #668