lsp-definition/lsp-references: fall back to grep if no server is configured
Many times you may be editing files for which you have not installed
a language server.  If you press `gd` (goto definition) with your
cursor on some text, kakoune will give an error with an info message
something like:

    "Language server is not configured for filetype `<something>`"

We can do better. As a fallback, run "grep" on the word found on the
cursor _and_ display the error as previously.
Same for references (`gr`).

Closes #690

Co-authored-by: Johannes Altmanninger <aclopte@gmail.com>