Preserve placeholders of existing completion snippets
Given this Rust program

	fn addition(a: i32, b: i32) -> i32 {
	    a + b
	}

Typing this(*)

	addi<c-n>addi<c-n>1<tab>2<tab>3<tab>

produces

	addition(addition(1, 2    3    %()), b)

but I expect

	addition(addition(1, 2), 3)%()

Fix this by preserving existing snippets when inserting new ones.

Note that range-specs are always sorted. To cope with this, assume that
new snippet placeholders are strictly left of existing placeholders,
which should be true in practice.

(*) With <tab> mapped to lsp-snippets-select-next-placeholders as
recommended in the README, typing