Use text edit also for completions that don't end exactly at the cursor
There is a very annoying race condition with rust-analyzer where we
sometimes get the expected

	lsp-snippets-insert-completion self.some_method(${1:param1}, ${2:param2})$0

but other times we get

	lsp-snippets-insert-completion self.some_method(…)$0

The second is because we believe we can't use
"completionItem.textEdit.insertText", so we fall back to
"completionItem.insertText" and "completionItem.label" as specified.

Our support for textEdit here is conditional on the fact that the text
edit's range (16..20) is strictly left of the cursor but that was 17
in my reproduction. Maybe stale buffer data?  Or this is triggered
by the "…" character, see https://github.com/kak-lsp/kak-lsp/pull/455

I haven't had time to investigate more; for now I don't think there
are any realistic downsides to using textEdit here, so let's do that.