Fix lsp-snippets-select-next-placeholders when a completion is selected
Given

	enum Enum {
	    Variant1,
	    Variant2,
	}
	fn takenum(_a: Enum, b: i32) -> i32 {
	    b
	}

and <tab> mapped to lsp-snippets-select-next-placeholders, typing

	taken<c-n>Enu<c-n><tab>

produces

	takenum(Enum::Variant1, )%()

but it should be

	takenum(Enum::Variant1, %())_

The bug comes about because lsp-snippets-select-next-placeholders
will be run again by the on-accept callback set in
lsp-snippets-insert-completion.
Change that one to only select a placeholder if we actually inserted
new snippets.
Also make sure that lsp-snippets-select-next-placeholders accepts any
selected completion before doing anything, else things would run in
the wrong order.