Discard requests from scratch buffers
If a scratch buffer is given an LSP-enabled filetype, we might end
up crashing as Url::from_file_path() will fail. I don't think there
is any use for scratch buffers yet so let's filter them out before
we have a chance to crash.

This was found on a file opener that uses prompt -on-change to
preview files.