Improve error reporting when jump.kak is missing
With the rename from grep_current_line to jump_current_line,
if jump.kak was not loaded we end up erroring like

	option not found: 'grep_current_line'

which is confusing because that's not the root cause.  Complain about
the root cause, but only if grep_current_line is missing because
otherwise Kakoune is probably too old to have jump.kak.

Also, besides reporting the error, fall back to declaring
jump_current_line ourselves, because most things work fine without
jump.kak, no need for a hard dependency.