Use more popular binary names vscode-{html,css,json}-language-server
It looks like upstream chose "vscode-*-languageserver" but apparently
upstream doesn't really service non-VSCode users - presumably that
was the motivation for the "vscode-languageservers-extracted" npm
package which seems to be using "vscode-*-language-server".  When we
added the config, Arch packages only had the upstream name but now
they offer both, so it looks like we should change our config to
follow the majority.

Closes #731