Insert the specified insertText for snippet-style completions,
We have a hack to fake support for textEdit.  Collateral damage of
that is that for completions that have `textEdit` and are snippets, we
use the wrong thing as insertText; since kakoune filters completions
based on that value, this means that in some cases valid completions
will be filtered away (if the current query is not a subsequence of
the completion.
Reproducible by completing after "div*3" with this config:

	[language_server.emmet-language-server]
	filetypes = ["html", "astro", "heex", "eex", "vue", "typescriptreact", "javascriptreact"]
	roots = ["package.json", ".git"]
	command = "emmet-language-server"
	args = ["--stdio"]

Disable the hack when we have a snippet, because we'll replace it
anyway when we expand the snippet as the completion is selected.