Move experimental client capabilities to the per-server config
rust-analyzer and the Dart language server have conflicting
interpretations of "experimental.commands".

I'm not aware of two servers that implement the same "experimental"
capability (though I didn't look much).

Move this capability to the server-specific config, so it's only sent
to the intended servers. Additionally, send it to all Rust servers,
as a temporary compatibility hack, so as to not force users to update
their config just yet.

Closes #758