Fix exit not being sent on lsp-stop
We send a request with `method="exit"` to all controllers with a
meta_for_session() meta, with no filename.  Commit 3a87b3d (Discard
requests from scratch buffers, 2024-01-15) broke the shutdown sequence,
so work around that here.

Technically, a scratch buffer with an empty name is allowed but I'll
worry about that later.