Write logs to the *debug* buffer
While at it, also add the lsp.* hooks to disabled_hooks
for scratch buffers.

This is hacky - we shouldn't be messing with the user's preference
but it seems to be the pragmatic approach. If we revisit this,
then we should make sure to filter out requests from scratch buffers
early in src/session.rs (unless it's "Exit::METHOD" or the upcoming
"kakoune/did-change-option"); at least the ones from hooks.  This is
because they add a lot of noise to *debug* when navigating scratch
buffers (the controller keeps restarting).