Clear snippet placeholders before inserting new placeholders in later position
When creating multiple sets of placeholders in insert mode
without visiting all of them with lsp-snippets-select-next-placeholders,
the behavior can be very confusing.

The intended use case for multiple concurrently active sets of
placeholders is nested function calls. Since placeholders are always
visited in lexical order, new placeholders are always inserted before
existing ones. if it's the other way round, the user must have moved
the cursor in insert mode. Let's clear snippets in that case.

Note that this commit assumes we have a single-selection only.
This is okay because complex snippets scenarios are rarely used
with multiple selections. Also there is another bug where the wrong
placeholders are selected for some selections.

Closes #770