Disable lsp hooks in buffers where the language server fails to start
Enabling LSP globally with lsp-enable will lead to noisy ERRO logs
when navigating files where the default language server is not
installed. Now that we log to the *debug* buffer this is
quite intrusive.

As a workaround, disable lsp hooks for such servers.  This is ugly
(we shouldn't be using disabled_hooks); hoping to find a better
solution eventually.