Remove lsp-if-no-servers hook ordering workaround
lsp-if-no-servers exists because we want the user to be able to
override servers using familiar hooks:

	eval %sh{kak-lsp --kakoune}
	hook global BufSetOption filetype=(?:c|cpp) %{
	    set-option buffer lsp_servers %{
	        [clangd]
	        root = "/path/to/project"
	    }
	}

without regard to hook ordering issues. The hook ordering issues should
only arise when the user's hook is defined before kak-lsp is loaded
(when the "eval %sh{kak-lsp --kakoune}" comes after the hook).

We can probably assume this doesn't happen, simplifying the default config.