Get rid of User hook, for a more copy-paste friendly default config
The LSPDefaultConfig hook exists to be able to delay
root detection etc. until after `lsp-enable`.
This has two reasons:
1. it also works for `kak main.c -e 'eval %sh{kak-lsp --kakoune}'`
2. no shell calls until `lsp-enable`

I realized that reason 1 is not so important because one can use
`kak -E`.  Reason 2 is probably not a big deal. In future we could
change root detection again to be declarative in most cases, removing
this reason.