Rework request sending
Consolidate editor-side request-sending in one central place.
Use "echo -to-file" instead of "kak-lsp --request" because that makes
it easier to share logic without creating a ton of shells.

This allows to
- consistently fail when lsp-enable* hasn't been used
- implement quoting correctly (for %val{buffile} etc) while still
  reducing the number of needed processes (shell, sed etc)
- get rid of commands like lsp-did-change-sync because there is no
  race with fifo semantics

And some more, see the following commits.

For simplicity, remove the feature to use a kak-lsp instance from
multiple Kakoune sessions. I'm not sure if this is useful; we can
always add it back.

Since trying to open the fifo will hang if there is no reader,
it's important that we always remove it when kak-lsp exits, so add
a signal handler.

Also make sure to exit on panic (after printing a backtrace), this
should be less confusing now.

Also make lsp_debug changes apply if using an externally-launched
kak-lsp.

While at it, implement a timestamp-checking logic that doesn't need
shell calls.  Thanks to Tobias Pisani for the idea.

Note that we still copy the buffile contents a few times,
but we can optimize this now.