Try to implement refresh requests
This kind of fixes codeLens, semanticTokens, inlayHints
not being shown with rust-analyzer until the first change to a file.
(Though for semanticTokens we had a workaround).

This fix currently only works for the first client that triggers a
NormalIdle hook.

Future work is to support workspace/inlineValue/refresh and
workspace/diagnostic/refresh (though it's unclear if that's useful
to us).

See https://github.com/rust-lang/rust-analyzer/issues/14048